RAG/requirements.txt
streamlit==1.38.0
chromadb==0.5.5
langchain==0.3.7
langchain-core==0.3.20
langchain-community==0.3.7
langchain-google-genai==2.0.2
google-genai==0.3.0
python-dotenv==1.0.1
Pillow==10.4.0
